Define EffectImpact by EffectType
Removes switch statement to determine EffectImpact but matches the
functionality. This commit does not change the ranking of EffectTypes or
the returned EffectImpact. In doing so, it highlights that the order and
comments about EffectImpacts do not match the implementation. Declaring
the EffectImpact in the enum constructor helps to prevent reintroducing
these discrepancies.