Synchronize everything in Transcript
A long-standing concurrency bug (pcingola/SnpEff#157) was causing
get5primeUtrs() to throw ConcurrentModificationException in
multithreaded mode. Running with large numbers of cores (96, for
instance) exacerbated the problem.

A Transcript was being shared between threads, and they were
concurrently modifying the utrs member. This change makes all methods
synchronized to prevent concurrent use of any of the internals. That
might be more synchronization than absolutely necessary, but it
appears to solve the problem without a dramatic performance penalty.